feat: Restrict cart access for guest users
This commit implements a check in `BaseActivity` to prevent guest users from accessing the cart. If a guest user clicks the cart button, a Toast message "You must be logged in to access the cart" is displayed.

A new method `isSignedAsGuest()` has been added to `AuthManager` to determine if the current user is anonymous.